rename batch-strategies and optimizers folders
